cec: move EnvoyL7LBBackendSyncer to package pkg/ciliumenvoyconfig
With the introduction of a dedicated k8s watcher for the CRDS
`CiliumEnvoyConfig` and `CiliumClusterwideEnvoyConfig`, the package
`pkg/ciliumenvoyconfig` is the only user of the `EnvoyL7LBBackendSyncer`.

Hence, let's move the syncer to this package and make it package private.
