+++ pkg/ciliumenvoyconfig/cell.go
@@ -28,6 +28,7 @@ var Cell = cell.Module(
 	"CiliumEnvoyConfig",
 
 	cell.Invoke(registerCECK8sWatcher),
+	cell.ProvidePrivate(newEnvoyServiceBackendSyncer),
 )
 
 type watchParams struct {
@@ -49,7 +50,7 @@ type watchParams struct {
 
 	Proxy         *proxy.Proxy
 	XdsServer     envoy.XDSServer
-	BackendSyncer *envoy.EnvoyServiceBackendSyncer
+	BackendSyncer *envoyServiceBackendSyncer
 
 	CECResources  resource.Resource[*ciliumv2.CiliumEnvoyConfig]
 	CCECResources resource.Resource[*ciliumv2.CiliumClusterwideEnvoyConfig]
