+++ pkg/ciliumenvoyconfig/envoy_l7lb_backend_syncer_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package envoy
+package ciliumenvoyconfig
 
 import (
 	"net/netip"
